pyproject.toml
[project]
name = "inf-502-final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dependency-injector>=4.48.2",
    "nicegui[highcharts]>=3.3.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygithub>=2.8.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.uv]
package = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
main_file = "app/infrastructure/web/main.py"

